+++ changelog.md
@@ -12,4 +12,4 @@ Refactored Cart to show count and sum for added items. Added some minimal stylin
 Added services with various registrations. Added About component with newly developed services. 
 Added a directive which changes border on click.
 
-
+Поправьте ошибки ng lint